@@ -3,36 +3,38 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@azure/msal-browser": "2.21.0",
-    "@azure/msal-react": "^1.2.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/react-fontawesome": "^0.1.11",
-    "@microsoft/applicationinsights-web": "^2.7.2",
+    "@azure/msal-browser": "2.37.1",
+    "@azure/msal-react": "1.5.8",
+    "@fortawesome/fontawesome-svg-core": "6.4.0",
+    "@fortawesome/free-solid-svg-icons": "6.4.0",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "@microsoft/applicationinsights-web": "^3.0.2",
     "@stripe/stripe-js": "^1.9.0",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "27.4.0",
-    "@types/jquery": "3.5.13",
-    "@types/node": "17.0.8",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-router-dom": "^5.3.2",
+    "@types/jest": "29.5.2",
+    "@types/jquery": "3.5.16",
+    "@types/node": "^16.18.34",
+    "@types/react": "18.2.9",
+    "@types/react-dom": "^18.2.4",
+    "@types/react-router-dom": "^5.3.3",
     "bootstrap": "^4.5.2",
-    "jquery": "^3.5.1",
-    "mini-css-extract-plugin": "2.4.5",
-    "node-sass": "^7.0.1",
+    "jquery": "^3.7.0",
+    "mini-css-extract-plugin": "2.7.6",
+    "node-sass": "^9.0.0",
     "popper.js": "^1.16.1",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.4.0",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^5.0.0",
-    "tslib": "^2.3.1",
-    "typescript": "4.5.4",
-    "use-async-effect": "^2.2.5"
+    "react": "^18.2.0",
+    "react-bootstrap": "2.7.4",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.12.0",
+    "react-scripts": "^5.0.1",
+    "tslib": "^2.5.3",
+    "use-async-effect": "^2.2.7"
+  },
+  "devDependencies": {
+    "typescript": "4.9.5",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3"
   },
   "scripts": {
     "start": "react-scripts start",
