@@ -1,20 +1,33 @@
-﻿import { useEffect } from 'react';
-import { withRouter } from 'react-router-dom';
-import { useHistory } from 'react-router-dom'
+﻿import React from 'react';
+import { useEffect } from 'react';
+import { useLocation, useNavigate, useParams } from 'react-router-dom';
 
 function ScrollToTop(props: any) {
-    const history = useHistory();
+    let location = useLocation();
 
     useEffect(() => {
-        const unlisten = history.listen(() => {
+        
             window.scrollTo(0, 0);
-        });
-        return () => {
-            unlisten();
-        }
-    }, []);
+        
+    }, [location]);
 
     return (null);
+}
+
+function withRouter(Component:any) {
+    function ComponentWithRouterProp(props:any) {
+        let location = useLocation();
+        let navigate = useNavigate();
+        let params = useParams();
+        return (
+            <Component
+                {...props}
+                router={{ location, navigate, params }}
+            />
+        );
+    }
+
+    return ComponentWithRouterProp;
 }
 
 export default withRouter(ScrollToTop);
